-Main file change to keep it more clean & added Window class
